+++ ced_vdop/ced_vdop.h
@@ -12,7 +12,7 @@
 #define CED_DEMO_08_H_298837492837
 
 struct SDrawCache {
-	::gpk::array_pod<::gpk::SCoord2<int32_t>>			PixelCoords				= {};
+	::gpk::array_pod<::gpk::SCoord2<int16_t>>			PixelCoords				= {};
 	::gpk::array_pod<::gpk::STriangleWeights<float>>	PixelVertexWeights		= {};
 	::gpk::array_pod<::gpk::SCoord3<float>>				LightPointsModel		= {};
 	::gpk::array_pod<::gpk::SColorBGRA>					LightColorsModel		= {};
